[Xperia Z5, Kitakami] msm: camera2: cpp: adjust poll time
The current poll value is pretty high in the order of ms. The
MICRO is capable of handling commands somewhat faster than current
delay between each poll read. This change optimize the latency
between each poll read to save 5 ms in the most optimal case.

The usleep values are chosen based on experiments
and this was the smallest number which works. This
sleep is needed to leave enough time for Microcontroller
to recover and work properly between each poll reads.

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>