[PORT][FPC1145] drivers: fingerprint: boost cpufreq on unlock
To reduce wake-up latency when using the fingerprint scanner we
register an input device and send a "key press" when the scanner
is being used during screen off. This will boost the cpus through
cpu-boost and reduce the screen on latency.

General idea "stole" from this commit sultanxda/android_kernel_oneplus_msm8996@f194585

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

this needs the commit

'display: add a simple api to query the display state (on/off) at any point in time'

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>

Conflicts:
	drivers/input/misc/fpc1145_tee.c