[PORT] net: wireless: bcmdhd: Fix suspend failure due to wakelock/wakeunlock mismatch
There is an case that shows wake lock remains high number
and prevent device go into sleep.
This is due to not calling unlock as same as lock called.
Fixed lock/unlock sequence to release wake lock correctly.

Bug: 23815756

Change-Id: I0eab6f1ad0d2fa5e65219ee7f4537e9c9371138b
Signed-off-by: Insun Song <isong@broadcom.com>
Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>

port the fix + add the lock where it was supposed to be ...

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>