DTS: Remove the core hotplug on low battery feature
By default BCL is configured to hotplug cpu4
and cpu5 when the SOC battery level threshold
(currently 10%) is reached.

This behaviour is unexpected and can result in
major performance drops.

So we reconfigure BCL to avoid hotplugging any
cores once the SOC battery level threshold
is reached.

We still have frequency mitigation for the same
cores (which doesn't make sense as without this
patch they are force-hotplugged anyways).

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>