net: wireless: bcmdhd: Fix del_station function parameters
This function had bad parameters for our current kernel.
Our 80211 implementation is a backport from 3.18 kernel, which
changes the del_station function signature in the struct
cfg80211_ops.

This also means that the function has never been working
correctly, so it was trying to delete an invalid mac address
resulting in unexpected behavior.