[PORT] msm: mdss: reduce pm runtime autosuspend delay to 50 ms
In the current implementation, the pm runtime autosuspend timeout is set
to 200 ms. This means that the MDSS idle power collapse would be delayed
by 200 ms after idle screen is detected. Though the autosuspend delay
can be controlled by userspace, it is still beneficial to reduce the
default value to 50 ms based on a recent analysis for use-cases
involving periodic single (or a small number of) frame update(s). This
should not affect normal use cases since the idle screen detection logic
itself waits for 6 VSYNCs for no frame updates, and should prevent
bouncing back and forth between runtime suspend and active states.

Change-Id: Ib562fd18dcdd66791e272ed2c37ff383621a8a23
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Anurag Singh <anursing@codeaurora.org>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>