sched: core_ctl: Don't consume big cores when screen is off
We never need too much cores when screen is off, so don't turn them on.
To achieve this, we add a hook to the cpu_notifier callback function in
core_ctl and tell it to use 0 for big_cluster's max cpu when screen is on.

Note that I use Franco's dispaly state API, so we should wait till this
callback function is used. From my test this callback is very quick (<1s),
so it is acceptable.

TODO:

1. Add tuneables caching to recover old parameters.
2. Make it a standalone driver.

Signed-off-by: Wang Han <416810799@qq.com>