slimbus: overclocking
Increase root freq from 11 (24.526MHz) to 31 max (27+MHz) which is the
largest it will take .Increase clock gear from 7 to 10 (max again).

This should cut down the audio driver <- --> audio chip latency
significantly, we might need to take a look at other components but this
should help quite a bit already with "Android Audio's 10 Millisecond
Problem".

Ultimately the "superpowered" mediaserver would be the solution with
apps that support it - but since we don't have it, we need to resort to
other tweaks for now ...

kudos to poondog and Francis

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>