msm: qpnp-power-on: Ignore ghost power key events
This happens to my old power key that it will report some events
in a samll period of time and that messes up everything.

To workaround this, we add a hook into qpnp-power-on driver and
record the time interval of different power key events, then reject
them based on hardcoded time interval.

This idea is taken from vendor change to cheeseburger.

Signed-off-by: Wang Han <416810799@qq.com>