BACKPORT:  Driver core: wakeup the parent device before trying probe
If the parent is still suspended when driver probe is
attempted, the result may be failure.

For example, if the parent is a PCI MFD device that has been
suspended when we try to probe our device, any register
reads will return 0xffffffff.

To fix the problem, making sure the parent is always awake
before attempting driver probe.

Bug: 62492269
Change-Id: I5580533b75d45d4c4b4f138a99ade7ac016f042c
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>