[reference] drivers/net: fix livelock issues
Preempt-RT runs into a live lock issue with the NETDEV_TX_LOCKED micro
optimization. The reason is that the softirq thread is rescheduling
itself on that return value. Depending on priorities it starts to
monoplize the CPU and livelock on UP systems.

Remove it.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
---
 drivers/net/ethernet/atheros/atl1c/atl1c_main.c      | 6 +-----
 drivers/net/ethernet/atheros/atl1e/atl1e_main.c      | 3 +--
 drivers/net/ethernet/chelsio/cxgb/sge.c              | 3 +--
 drivers/net/ethernet/neterion/s2io.c                 | 7 +------
 drivers/net/ethernet/oki-semi/pch_gbe/pch_gbe_main.c | 6 ++----
 drivers/net/ethernet/tehuti/tehuti.c                 | 9 ++-------
 drivers/net/rionet.c                                 | 6 +-----
 7 files changed, 9 insertions(+), 31 deletions(-)

we're not using any of these drivers but let's add it anyway,
might use it as reference for fixup of other drivers

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>