[reference] drivers/net: vortex fix locking issues
Argh, cut and paste wasn't enough...

Use this patch instead.  It needs an irq disable.  But, believe it or
not,
on SMP this is actually better.  If the irq is shared (as it is in
Mark's
case), we don't stop the irq of other devices from being handled on
another CPU (unfortunately for Mark, he pinned all interrupts to one
CPU).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

 drivers/net/ethernet/3com/3c59x.c |    8 ++++----
 1 file changed, 4 insertions(+), 4 deletions(-)

also use this one as reference, we're not using it anyway ...

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>