sched: Postpone actual migration disalbe to schedule
The migrate_disable() can cause a bit of a overhead to the RT kernel,
as changing the affinity is expensive to do at every lock encountered.
As a running task can not migrate, the actual disabling of migration
does not need to occur until the task is about to schedule out.

In most cases, a task that disables migration will enable it before
it schedules making this change improve performance tremendously.

[ Frank Rowand: UP compile fix ]

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Clark Williams <williams@redhat.com>
Link: http://lkml.kernel.org/r/20110927124422.779693167@goodmis.org
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
---
 kernel/sched/core.c | 251
+++++++++++++++++++++++++++-------------------------
 1 file changed, 132 insertions(+), 119 deletions(-)

+ assume "RT" on by default

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>

Conflicts:
	kernel/sched/core.c

due to non-applied 'sched: Generic migrate_disable' and subsequent
changes

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>