msm: qpnp-power-on: Make debounce time for power key configurable
This is a workaround for races in s3320 driver. The real issue is
there is somehow race conditions that when one turns screen on and
off too quickly, he will sometimes get a weird screen display.

As the issue is hard to address, don't let the user press the power
key too quickly. I think 0.75s is a good balance.

Signed-off-by: Wang Han <416810799@qq.com>

Conflicts:
	drivers/platform/msm/qpnp-power-on.c

due to missing

(1+ additions)

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>