arch/arm/boot/dts/qcom/msm8994.dtsi: slub: enable sanity checks, redzoning & poisoning in addition to nomerge (is automatically implied but let's be super pedantic and avert any regressions and issues)
Description:

slub_debug=FZP
Enables sanity checks (F), redzoning (Z) and poisoning (P). Sanity
checks are self-evident and come with a modest performance impact, but
this is unlikely to be significant on an average Tails system. The
checks are basic but are still useful both for security and as a
debugging measure. Redzoning adds extra areas around slabs that detect
when a slab is overwritten past its real size, which can help detect
overflows. Its performance impact is negligible.

Poisoning writes an arbitrary value to freed objects, so any
modification or reference to that object after being freed or before
being initialized will be detected and prevented. This prevents many
types of use-after-free vulnerabilities at little performance cost.

An additional note: any time slub_debug= is put in the kernel command
line,  slab_nomerge is implied. But having slab_nomerge explicitely
declared can help prevent regressions where disabling of debugging
features is desired but re-enabling of merging is not.

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>